requirements/base.txt
pytz==2018.4  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==3.3.1  # https://github.com/evansd/whitenoise
redis>=2.10.5  # https://github.com/antirez/redis
celery==3.1.25  # pyup: <4.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==2.0.5  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis
django-cuser==2017.3.16  # https://github.com/Alir3z4/django-cuser
django-grappelli==2.11.1  # https://github.com/sehmaschine/django-grappelli
django-nested-admin==3.0.21  # https://github.com/theatlantic/django-nested-admin
django-tables2==1.21.2  # https://github.com/jieter/django-tables2

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==1.1.0
djangorestframework-datatables==0.3.0
django-rest-swagger==2.2.0

